+++ sdk/tests/iot/hub/test_az_iot_hub_client_telemetry.c
@@ -170,8 +170,7 @@ static void test_az_iot_hub_client_telemetry_get_publish_topic_with_props_unfill
       = az_span_create(test_prop_unfilled_buf, sizeof(test_prop_unfilled_buf));
   assert_int_equal(az_iot_message_properties_init(&props, test_prop_unfilled_span, 0), AZ_OK);
   assert_int_equal(
-      az_iot_message_properties_append(
-          &props, AZ_SPAN_FROM_STR("key"), AZ_SPAN_FROM_STR("value")),
+      az_iot_message_properties_append(&props, AZ_SPAN_FROM_STR("key"), AZ_SPAN_FROM_STR("value")),
       AZ_OK);
   assert_int_equal(
       az_iot_message_properties_append(
@@ -213,7 +212,7 @@ test_az_iot_hub_client_telemetry_get_publish_topic_with_options_with_props_small
   assert_true(
       az_iot_hub_client_telemetry_get_publish_topic(
           &client, &props, test_buf, sizeof(test_buf), &test_length)
-      == AZ_ERROR_INSUFFICIENT_SPAN_SIZE);
+      == AZ_ERROR_NOT_ENOUGH_SPACE);
 }
 
 static void test_az_iot_hub_client_telemetry_get_publish_topic_no_options_with_props_succeed(
@@ -261,7 +260,7 @@ test_az_iot_hub_client_telemetry_get_publish_topic_no_options_with_props_small_b
   assert_true(
       az_iot_hub_client_telemetry_get_publish_topic(
           &client, &props, test_buf, sizeof(test_buf), &test_length)
-      == AZ_ERROR_INSUFFICIENT_SPAN_SIZE);
+      == AZ_ERROR_NOT_ENOUGH_SPACE);
 }
 
 static void
@@ -316,7 +315,7 @@ test_az_iot_hub_client_telemetry_get_publish_topic_with_options_module_id_with_p
   assert_true(
       az_iot_hub_client_telemetry_get_publish_topic(
           &client, &props, test_buf, sizeof(test_buf), &test_length)
-      == AZ_ERROR_INSUFFICIENT_SPAN_SIZE);
+      == AZ_ERROR_NOT_ENOUGH_SPACE);
 }
 
 int test_az_iot_hub_client_telemetry()